pyproject.toml
[tool.poetry.dependencies]
python-telegram-bot = "^20.0"